New workflow for TRSV and SYMMV, CSR optimize, and symmetric Gauss-Seidel.
optimize hints are now stored in a linked list inside the matrix structure. The optimize function then checks all the hints to decide which optimization to perform:
if only MV hints have been passed, optimize for MV, otherwise store a clean sorted CSR matrix with some additional info.
To clean CSR:
- check that the input matrix is valid (no indices out of bounds)
- check that the matrix is sorted
- check that the diagonal is full
if not:
create a copy of the matrix with sorted elements and artificial 0 fill-ins on the missing elements of the diagonal.
Removed hint_id from the main matrix structure as it's not used anymore.

Addition of trsv solvers for all configurations in the inspector executor framework. the 4 double versions (lower/upper triangle transpose or not) have an optimized avx version.
The user facing interfaces aoclsparse_d/strsv() use the descriptor to decide on which part of the matrix.

Similarly, symmetric matrix vector products are added for use with aoclsparse_d/smv(). The matrix vector product dispatcher has been slightly modified to accomodate for more type of matrices.

Simple tests are added for the hinting and trsv functionality in a new folder under tests.

CG direct interface now makes calls directly to csrmv and Gauss-Seidel preconditioner has been added.

A new flag in aocl_sparse_matrix has been created to safely be able to call destroy_2m().

Signed-off-by: bmarteau <<bmarteau@amd.com>>
AMD-Internal: [CPUPL-2233]
Change-Id: If5d834a58d948e5f6218b107379f5f3d302da783