Build doc ONLY if explicitly requested
This patch removes the ALL flag from the docs target.
The idea is when -DBUILD_DOCS=On, then the docs should only be built
once and not every time `make all` is invoked. To build the docs
(and bump the version) invoke `make docs`.

Rationale:
This change allows a developer to build mutiple times the library
without having to execute doxygen. Once the source changes are
complete, then the doc can be updated by calling `make docs`.

README.md updated explaining how to build the documentation.

Signed-off-by: andrsajo <andrew.sajo@amd.com>

AMD-Internal: N/A
Change-Id: Ic432e25a0e111b6c5b03c747ad3a53e5289177e4