Add `tools` directory to AOCL-Sparse
This commit adds a home to house all tools required for development, QC,
etc. It also adds a helper script to identify all existing warnings and
produce a summary table, similar to

 Statistics:

   count warning
    245 -Wunused-parameter
     79 -Wunused-variable
      7 -Wunused-but-set-variable
      6 -Wtautological-compare
      5 -Wmaybe-uninitialized
      4 -Wrange-loop-construct
      4 -Wimplicit-fallthrough=

Note: this folder does not have to be shipped to the public repository
or in any package. It is only for internal use.

Signed-off-by: Andrew Sajo <andrew.sajo@amd.com>

AMD-Internal: N/A
Change-Id: I19c7a0cd4c25ea55fae63e3515f2124dcc504068