Two small tools to check/fix copyright message.
check_copyright.sh looks for the current year in the copyright message
and if not found, the file is reported.

fix_copyright.sh attempts to fix it, it might be a bit flaky as it expects
the copyright message to be pretty exact but if it is, it will modify it
as needed.

Signed-off-by: janfiala
Change-Id: Ia8f1acafea62af3953da2ce055ce1b849f82114e