Allow partial sorting (L/D/U) in 'Clean CSR' format for hinted SPMV/TRSV
To be able to call TRSV only on a portion of a general matrix (e.g.,
solve L triangle of the matrix) we set up 'Clean CSR' notion in 4.0
in the sense that we can point to the diagonal element and work
only with the submatrix without testing every element for being
in L or D or U.

In 4.0 we required complete sorting in rows of CSR otherwise a sorted
copy of the matrix would have been produced. This is not necessary and
only partial sorting is sufficient, i.e., elements of L are followed
by D and then by U, but sorting in each of the group is not required.

That's allowed by this change.
Plus added deallocation in functionality tests so that they come clean
in ASAN.

Signed-off-by: janfiala
AMD-Internal: CPUPL-2232
Change-Id: Icfe2e9f895520bfbe1845b7bf0be449913da2265