Pick up correct version of Blis/Flame LP64/ILP64
If Blis and Flame from monthly or public drop are used,
the directory structure is different and instead of
lib/ILP64 and include/ILP64, lib_ILP64 and include_ILP64
should be seeked instead. In that case one should avoid
the issue of picking up libs and includes from lib/ and
include/ which can point to either version.

Added static_assert to check at compile time if the integer
types are compatible.

In some weird cases, aoclsparse_bench was picking up
aoclspare.h header from the installed libraries instead of
the local one. The order of BLIS and Sparse headers need to
be explicit.

Signed-off-by: janfiala
Change-Id: Ica110289e7824ef81668f1e1d352c348f9533016