Modify benchmark error handling
Till now we used macro CHECK_AOCLSPARSE_ERROR() to check if any
aoclsparse call during benchmarking hasn't failed and in that case
we would exit immediately. This is not suitable if we want to
run multiple tests or benchmark against an external library.

With the new approach, the new macro rather throws a unique
exception which we catch during the benchmarking call and
clean up. The failure is indicated to ctest through nonzero
return from the main function.

Signed-off-by: janfiala
AMD-Internal: CPUPL-4409
Change-Id: I08d799eaa112cc16019af4fccab8145f520ed649