Skip bench stats computation if iters==1
Avoid exception from boost/math/statistics when external benchmarking
is turned on and aoclsparse-bench runs only with 1 iteration in the
hot loop (--iters==1). In that case statistics cannot be computed
due to the insufficient degrees of freedom, the output is adjusted
reflecting the uncertainity of the comparison result.

Signed-off-by: janfiala
AMD-Internal: CPUPL-4409
Change-Id: Ie6b0e65929e0bd606cad1cf14465a16b76830223