Refactor CSC matrix handling
CSC were already stored as aoclsparse::csr matrix class, however, their
dimensions were considered special and additional checks had to be done
to determine if the matrix is CSR and thus m x n or CSC with n x m dimensions.
This addresses it - CSC matrix is considered as ordinary CSR and its DOID indicates
that it is transposed. The dimensions are natural (i.e., the dimension of mat->ptr[]
array is always mat->m+1).

* Updated CSR constructor to always allocate memory for ptr with size m.
* Removed type check for csc_mat during ptr memory allocation.
* Added optional doid parameter to CSR constructor (default: doid::gn).
* Updated create_csc to create CSR matrix with doid::gt and dimensions (n, m).
* Removed aoclsparse_csc_mat type; now use A->mats[0]->doid::gt to identify CSC matrices.
* Updated all usages of aoclsparse_csc_mat accordingly.
* Bugfix: correctly propagate base of the matrix in CSRMM
   1-base input matrix was incorrectly considered as 0-base
   when searching for a matching DOID matrix, assuming that
   it is always internally generated which would be 0-based.

AMD-Internal: CPUPL-7072